requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
fastmcp>=0.1.0
